7/11 complete!

+++ ItemProperties.cs
@@ -3,7 +3,7 @@
     public class ItemProperties //or as public interface
     {
         public string Name { get; set; }
-        public double Price { get; set; }
+        public decimal Price { get; set; }
         public string Description { get; set; }
         public string ItemNumber { get; set; } //keep as string, or use int for multiple menus/txt files ; add item # for easy user selection?
         //public int Quantity { get; set; }
@@ -13,7 +13,7 @@
 
         }
 
-        public ItemProperties(string category, string name, double price, string description) //, int quantity
+        public ItemProperties(string category, string name, decimal price, string description) //, int quantity
         {
             Name = name;
             Price = price;
