+++ MenuFilesIO.cs
@@ -26,7 +26,7 @@ namespace CashRegApp
                 while (line != null)
                 {
                     string[] itemproperties = line.Split("|");
-                    itemProperties.Add(new ItemProperties (itemproperties[0], itemproperties[1], double.Parse(itemproperties[2]), itemproperties[3])); //, int.Parse(itemproperties[4])
+                    itemProperties.Add(new ItemProperties (itemproperties[0], itemproperties[1], decimal.Parse(itemproperties[2]), itemproperties[3])); //, int.Parse(itemproperties[4])
                     line = reader.ReadLine();
                 }
